// src/types/user.ts
// {
// "address": {
// "geolocation": {
// "lat": "-37.3159",
// "long": "81.1496"
// },
// "city": "kilcoole",
// "street": "Lovers Ln",
// "number": 7267,
// "zipcode": "12926-3874"
// },
// "id": 2,
// "email": "morrison@gmail.com",
// "username": "mor_2314",
// "password": "83r5^_",
// "name": {
// "firstname": "david",
// "lastname": "morrison"
// },
// "phone": "1-570-236-7033",
// "__v": 0
// },

type User = {
  id?: number;
  name?: {
    firstname?: string;
    lastname?: string;
  };
  phone?: string;
  address?: {
    geolocation?: {
      lat: string;
      long: string;
    };
    city?: string;
    street?: string;
    number?: number;
    zipcode?: string;
  };
  username: string;
  email?: string;
  password: string;
};

export { User };